Add local stories linking feature
Stories in the same Project can now be referenced on another story's
description. A new parser was added on the StoryView to intercept and
turn '#<story_id>' into links which would later turn into StoryLinkViews
providing the current state of the story.

Add popover to story links

Fix and add new tests